@@ -1,11 +1,10 @@
-﻿namespace Yuksi.Domain.Abstractions
+﻿namespace Yuksi.Domain.Abstractions;
+
+public abstract class Entity
 {
-    public abstract class Entity
+    public Guid Id { get; set; }
+    protected Entity()
     {
-        public Guid Id { get; set; }
-        protected Entity()
-        {
-            Id = Guid.NewGuid();
-        }
+        Id = Guid.NewGuid();
     }
 }
