@@ -4,23 +4,24 @@ using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore;
 using Yuksi.Domain.Entities;
 
-namespace Yuksi.Infrastructure.Context
+namespace Yuksi.Infrastructure.Context;
+
+internal sealed class ApplicationDbContext : IdentityDbContext<AppUser, IdentityRole<Guid>, Guid>, IUnitOfWork
 {
-    internal sealed class ApplicationDbContext : IdentityDbContext<AppUser, IdentityRole<Guid>, Guid>, IUnitOfWork
+    public ApplicationDbContext(DbContextOptions options) : base(options)
     {
-        public ApplicationDbContext(DbContextOptions options) : base(options)
-        {
-        }
+    }
 
-        protected override void OnModelCreating(ModelBuilder builder)
-        {
-            builder.ApplyConfigurationsFromAssembly(typeof(DependencyInjection).Assembly);
+    protected override void OnModelCreating(ModelBuilder builder)
+    {
+        builder.ApplyConfigurationsFromAssembly(typeof(DependencyInjection).Assembly);
 
-            builder.Ignore<IdentityUserLogin<Guid>>();
-            builder.Ignore<IdentityRoleClaim<Guid>>();
-            builder.Ignore<IdentityUserToken<Guid>>();
-            builder.Ignore<IdentityUserRole<Guid>>();
-            builder.Ignore<IdentityUserClaim<Guid>>();
-        }
+        builder.Ignore<IdentityPasskeyData>();
+        builder.Ignore<IdentityUserPasskey<Guid>>();
+        builder.Ignore<IdentityUserLogin<Guid>>();
+        builder.Ignore<IdentityRoleClaim<Guid>>();
+        builder.Ignore<IdentityUserToken<Guid>>();
+        builder.Ignore<IdentityUserRole<Guid>>();
+        builder.Ignore<IdentityUserClaim<Guid>>();
     }
 }
