@@ -3,59 +3,69 @@ using Microsoft.AspNetCore.Identity;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
+using Npgsql;
 using Scrutor;
 using System.Reflection;
 using Yuksi.Domain.Entities;
+using Yuksi.Domain.Enums;
 using Yuksi.Infrastructure.Context;
 using Yuksi.Infrastructure.Options;
 
-namespace Yuksi.Infrastructure
+namespace Yuksi.Infrastructure;
+
+public static class DependencyInjection
 {
-    public static class DependencyInjection
+    public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
     {
-        public static IServiceCollection AddInfrastructure(this IServiceCollection services, IConfiguration configuration)
+
+        var connectionString = configuration.GetConnectionString("DefaultConnection");
+
+        var dataSourceBuilder = new NpgsqlDataSourceBuilder(connectionString);
+        dataSourceBuilder.MapEnum<ContactMessageStatus>();
+
+        var dataSource = dataSourceBuilder.Build();
+
+        services.AddDbContext<ApplicationDbContext>(options =>
         {
-            services.AddDbContext<ApplicationDbContext>(options =>
+            options.UseNpgsql(dataSource).UseSnakeCaseNamingConvention();
+        });
+
+
+        services.AddScoped<IUnitOfWork>(srv => srv.GetRequiredService<ApplicationDbContext>());
+
+        services
+            .AddIdentity<AppUser, IdentityRole<Guid>>(cfr =>
             {
-                options.UseSqlServer(configuration.GetConnectionString("SqlServer"));
-            });
+                cfr.Password.RequiredLength = 1;
+                cfr.Password.RequireNonAlphanumeric = false;
+                cfr.Password.RequireUppercase = false;
+                cfr.Password.RequireLowercase = false;
+                cfr.Password.RequireDigit = false;
+                cfr.SignIn.RequireConfirmedEmail = true;
+                cfr.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(5);
+                cfr.Lockout.MaxFailedAccessAttempts = 3;
+                cfr.Lockout.AllowedForNewUsers = true;
+            })
+            .AddEntityFrameworkStores<ApplicationDbContext>()
+            .AddDefaultTokenProviders();
 
-            services.AddScoped<IUnitOfWork>(srv => srv.GetRequiredService<ApplicationDbContext>());
+        services.Configure<JwtOptions>(configuration.GetSection("Jwt"));
+        services.ConfigureOptions<JwtTokenOptionsSetup>();
+        services.AddAuthentication()
+            .AddJwtBearer();
+        services.AddAuthorizationBuilder();
 
-            services
-                .AddIdentity<AppUser, IdentityRole<Guid>>(cfr =>
-                {
-                    cfr.Password.RequiredLength = 1;
-                    cfr.Password.RequireNonAlphanumeric = false;
-                    cfr.Password.RequireUppercase = false;
-                    cfr.Password.RequireLowercase = false;
-                    cfr.Password.RequireDigit = false;
-                    cfr.SignIn.RequireConfirmedEmail = true;
-                    cfr.Lockout.DefaultLockoutTimeSpan = TimeSpan.FromMinutes(5);
-                    cfr.Lockout.MaxFailedAccessAttempts = 3;
-                    cfr.Lockout.AllowedForNewUsers = true;
-                })
-                .AddEntityFrameworkStores<ApplicationDbContext>()
-                .AddDefaultTokenProviders();
+        services.Scan(action =>
+        {
+            action
+            .FromAssemblies(Assembly.GetExecutingAssembly())
+            .AddClasses(publicOnly: false)
+            .UsingRegistrationStrategy(RegistrationStrategy.Skip)
+            .AsMatchingInterface()
+            .AsImplementedInterfaces()
+            .WithScopedLifetime();
+        });
 
-            services.Configure<JwtOptions>(configuration.GetSection("Jwt"));
-            services.ConfigureOptions<JwtTokenOptionsSetup>();
-            services.AddAuthentication()
-                .AddJwtBearer();
-            services.AddAuthorizationBuilder();
-
-            services.Scan(action =>
-            {
-                action
-                .FromAssemblies(Assembly.GetExecutingAssembly())
-                .AddClasses(publicOnly: false)
-                .UsingRegistrationStrategy(RegistrationStrategy.Skip)
-                .AsMatchingInterface()
-                .AsImplementedInterfaces()
-                .WithScopedLifetime();
-            });
-
-            return services;
-        }
+        return services;
     }
 }
