@@ -1,29 +1,28 @@
 ﻿using Microsoft.AspNetCore.Identity;
 using Yuksi.Domain.Entities;
 
-namespace Yuksi.WebAPI.Middlewares
+namespace Yuksi.WebAPI.Middlewares;
+
+public static class ExtensionsMiddleware
 {
-    public static class ExtensionsMiddleware
+    public static void CreateFirstUser(WebApplication app)
     {
-        public static void CreateFirstUser(WebApplication app)
+        using (var scoped = app.Services.CreateScope())
         {
-            using (var scoped = app.Services.CreateScope())
+            var userManager = scoped.ServiceProvider.GetRequiredService<UserManager<AppUser>>();
+
+            if (!userManager.Users.Any(p => p.UserName == "admin"))
             {
-                var userManager = scoped.ServiceProvider.GetRequiredService<UserManager<AppUser>>();
-
-                if (!userManager.Users.Any(p => p.UserName == "admin"))
+                AppUser user = new()
                 {
-                    AppUser user = new()
-                    {
-                        UserName = "admin",
-                        Email = "admin@admin.com",
-                        FirstName = "Taner",
-                        LastName = "Saydam",
-                        EmailConfirmed = true
-                    };
+                    UserName = "admin",
+                    Email = "admin@admin.com",
+                    FirstName = "Taner",
+                    LastName = "Saydam",
+                    EmailConfirmed = true
+                };
 
-                    userManager.CreateAsync(user, "1").Wait();
-                }
+                userManager.CreateAsync(user, "1").Wait();
             }
         }
     }
